Twitter-API/Just Twitter Api/With library/requirements.txt
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
idna==2.8
oauthlib==3.1.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.7
